Struct compact_map::CompactMap

pub struct CompactMap<K, V, const N: usize> { /* private fields */ }

A map that inlines entries to avoid heap allocations for small maps.

Implementations

impl<K, V, const N: usize> CompactMap<K, V, N>

pub const fn new() -> Self

Creates an empty CompactMap.

The compact map will be able to hold up to N entries without spilling to the heap.

Examples

use compact_map::CompactMap;
let mut map: CompactMap<&str, i32, 16> = CompactMap::new();

pub const fn spilled(&self) -> bool

Returns true if the data has spilled into an std HashMap.

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 2>  = CompactMap::new();
assert!(!map.spilled());

map.insert(1, 2);
map.insert(3, 4);
map.insert(5, 6);
assert!(map.spilled());

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

When spilled, this number is a lower bound; the CompactMap<K, V> might be able to hold more, but is guaranteed to be able to hold at least this many.

Examples

use compact_map::CompactMap;
let map: CompactMap<i32, i32, 100> = CompactMap::new();
assert!(map.capacity() >= 100);

pub fn keys(&self) -> Keys<'_, K, V, N>

An iterator visiting all keys in arbitrary order. The iterator element type is &'a K.

Examples

use compact_map::CompactMap;

let map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

for key in map.keys() {
    println!("{key}");
}

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per docs in HashMap::keys, iterating over keys takes O(capacity) time.

pub fn into_keys(self) -> IntoKeys<K, V, N>

Creates a consuming iterator visiting all the keys in arbitrary order. The map cannot be used after calling this. The iterator element type is K.

Examples

use compact_map::CompactMap;

let map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

let mut vec: Vec<&str> = map.into_keys().collect();
// The `IntoKeys` iterator produces keys in arbitrary order, so the
// keys must be sorted to test them against a sorted array.
vec.sort_unstable();
assert_eq!(vec, ["a", "b", "c"]);

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn values(&self) -> Values<'_, K, V, N>

An iterator visiting all values in arbitrary order. The iterator element type is &'a V.

Examples

use compact_map::CompactMap;

let map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

for val in map.values() {
    println!("{val}");
}

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V, N>

An iterator visiting all values mutably in arbitrary order. The iterator element type is &'a mut V.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

for val in map.values_mut() {
    *val = *val + 10;
}

for val in map.values() {
    println!("{val}");
}

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn into_values(self) -> IntoValues<K, V, N>

Creates a consuming iterator visiting all the values in arbitrary order. The map cannot be used after calling this. The iterator element type is V.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

let mut vec: Vec<i32> = map.into_values().collect();
// The `IntoValues` iterator produces values in arbitrary order, so
// the values must be sorted to test them against a sorted array.
vec.sort_unstable();
assert_eq!(vec, [1, 2, 3]);

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn iter(&self) -> Iter<'_, K, V, N>

An iterator visiting all key-value pairs in arbitrary order. The iterator element type is (&'a K, &'a V).

Examples

use compact_map::CompactMap;

let map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

for (key, val) in map.iter() {
    println!("key: {key} val: {val}");
}

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V, N>

An iterator visiting all key-value pairs in arbitrary order, with mutable references to the values. The iterator element type is (&'a K, &'a mut V).

Examples

use compact_map::CompactMap;

let mut map: CompactMap<&str, i32, 3> = CompactMap::from([
    ("a", 1),
    ("b", 2),
    ("c", 3),
]);

// Update all values
for (_, val) in map.iter_mut() {
    *val *= 2;
}

for (key, val) in &map {
    println!("key: {key} val: {val}");
}

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn len(&self) -> usize

Returns the number of elements in the map.

Examples

use compact_map::CompactMap;

let mut a = CompactMap::default();
assert_eq!(a.len(), 0);
a.insert(1, "a");
assert_eq!(a.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

use compact_map::CompactMap;

let mut a = CompactMap::default();
assert!(a.is_empty());
a.insert(1, "a");
assert!(!a.is_empty());

pub fn drain(&mut self) -> Drain<'_, K, V, N>

Clears the map, returning all key-value pairs as an iterator. Keeps the allocated memory for reuse.

If the returned iterator is dropped before being fully consumed, it drops the remaining key-value pairs. The returned iterator keeps a mutable borrow on the map to optimize its implementation.

Examples

use compact_map::CompactMap;

let mut a = CompactMap::default();
a.insert(1, "a");
a.insert(2, "b");

for (k, v) in a.drain().take(1) {
    assert!(k == 1 || k == 2);
    assert!(v == "a" || v == "b");
}

assert!(a.is_empty());

pub fn extract_if<F>(&mut self, pred: F) -> ExtractIf<'_, K, V, F, N>
where F: FnMut(&K, &mut V) -> bool,

Available on crate feature extract_if only.

Creates an iterator which uses a closure to determine if an element should be removed.

If the closure returns true, the element is removed from the map and yielded. If the closure returns false, or panics, the element remains in the map and will not be yielded.

Note that extract_if lets you mutate every value in the filter closure, regardless of whether you choose to keep or remove it.

If the returned ExtractIf is not exhausted, e.g. because it is dropped without iterating or the iteration short-circuits, then the remaining elements will be retained. Use retain with a negated predicate if you do not need the returned iterator.

Examples

Splitting a map into even and odd keys, reusing the original map:

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 8> = (0..8).map(|x| (x, x)).collect();
let extracted: CompactMap<i32, i32, 8> = map.extract_if(|k, _v| k % 2 == 0).collect();

let mut evens = extracted.keys().copied().collect::<Vec<_>>();
let mut odds = map.keys().copied().collect::<Vec<_>>();
evens.sort();
odds.sort();

assert_eq!(evens, vec![0, 2, 4, 6]);
assert_eq!(odds, vec![1, 3, 5, 7]);

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&K, &mut V) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all pairs (k, v) for which f(&k, &mut v) returns false. The elements are visited in unsorted (and unspecified) order.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 16> = (0..8).map(|x| (x, x*10)).collect();
map.retain(|&k, _| k % 2 == 0);
assert_eq!(map.len(), 4);

Performance

When heapless: iterating over keys takes O(len) time.

When spilled: as per in std docs, iterating over keys takes O(capacity) time.

pub fn clear(&mut self)

Clears the map, removing all key-value pairs. Keeps the allocated memory for reuse.

Examples

use compact_map::CompactMap;

let mut a = CompactMap::default();
a.insert(1, "a");
a.clear();
assert!(a.is_empty());

impl<K, V, const N: usize> CompactMap<K, V, N>
where K: Eq + Hash,

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the CompactMap. The collection may reserve more space to speculatively avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

If current variant is heapless and self.len() + additional is greater than N, the map will spill to HashMap immediately; otherwise, it’s a no-op.

Panics

Panics if the new allocation size overflows usize.

Examples

use compact_map::CompactMap;
let mut map: CompactMap<&str, i32, 16> = CompactMap::new();
map.reserve(32);
assert!(map.capacity() >= 32);
assert!(map.spilled());

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

Tries to reserve capacity for at least additional more elements to be inserted in the HashMap. The collection may reserve more space to speculatively avoid frequent reallocations. After calling try_reserve, capacity will be greater than or equal to self.len() + additional if it returns Ok(()). Does nothing if capacity is already sufficient.

If current variant is heapless and self.len() + additional is greater than N, the map will spill to HashMap immediately; otherwise, it’s a no-op.

Errors

If the capacity overflows, or the allocator reports a failure, then an error is returned.

Examples

use compact_map::CompactMap;
let mut map: CompactMap<&str, i32, 16> = CompactMap::new();

map.try_reserve(10).expect("why is the test harness OOMing on a handful of bytes?");

pub fn spill(&mut self)

Manually spills to a HashMap.

Examples

use compact_map::CompactMap;
let mut map: CompactMap<i32, i32, 16> = CompactMap::new();
map.spill();
assert!(map.spilled());

pub fn shrink_into_heapless<const M: usize>( self ) -> Result<CompactMap<K, V, M>, CompactMap<K, V, N>>

Shrinks the map into a heapless map with capacity M.

Errors

If M is less than the current length of the map, the original map is returned.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 16> = CompactMap::new();
map.insert(1, 2);
map.insert(3, 4);
let map = map.shrink_into_heapless::<2>().unwrap();

pub fn shrink_to_fit(&mut self)

This is a proxy to the underlying HashMap::shrink_to_fit method. And it’s a no-op if the map is heapless.

Shrinks the capacity of the map as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 10> = CompactMap::new();
map.reserve(90);
map.insert(1, 2);
map.insert(3, 4);
assert!(map.capacity() >= 100);
map.shrink_to_fit();
assert!(map.capacity() >= 2);

pub fn shrink_to(&mut self, min_capacity: usize)

This is a proxy to the underlying HashMap::shrink_to method. And it’s a no-op if the map is heapless.

Shrinks the capacity of the map with a lower limit. It will drop down no lower than the supplied limit while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

If the current capacity is less than the lower limit, this is a no-op.

Examples

use compact_map::CompactMap;

let mut map: CompactMap<i32, i32, 10> = CompactMap::new();
map.reserve(90);
map.insert(1, 2);
map.insert(3, 4);
assert!(map.capacity() >= 100);
map.shrink_to(10);
assert!(map.capacity() >= 10);
map.shrink_to(0);
assert!(map.capacity() >= 2);

pub fn entry(&mut self, key: K) -> Entry<'_, K, V, N>

Gets the given key’s corresponding entry in the map for in-place manipulation.

Examples

use compact_map::CompactMap;

let mut letters = CompactMap::default();

for ch in "a short treatise on fungi".chars() {
    letters.entry(ch).and_modify(|counter| *counter += 1).or_insert(1);
}

assert_eq!(letters[&'s'], 2);
assert_eq!(letters[&'t'], 3);
assert_eq!(letters[&'u'], 1);
assert_eq!(letters.get(&'y'), None);

pub fn get<Q>(&self, k: &Q) -> Option<&V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.get(&2), None);

pub fn get_key_value<Q>(&self, k: &Q) -> Option<(&K, &V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns the key-value pair corresponding to the supplied key.

The supplied key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
assert_eq!(map.get_key_value(&1), Some((&1, &"a")));
assert_eq!(map.get_key_value(&2), None);

pub fn get_many_mut<Q, const M: usize>( &mut self, ks: [&Q; M] ) -> Option<[&mut V; M]>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Available on crate feature many_mut only.

Attempts to get mutable references to N values in the map at once.

Returns an array of length N with the results of each query. For soundness, at most one mutable reference will be returned to any value. None will be returned if any of the keys are duplicates or missing.

Examples

use compact_map::CompactMap;

let mut libraries = CompactMap::default();
libraries.insert("Bodleian Library".to_string(), 1602);
libraries.insert("Athenæum".to_string(), 1807);
libraries.insert("Herzogin-Anna-Amalia-Bibliothek".to_string(), 1691);
libraries.insert("Library of Congress".to_string(), 1800);

let got = libraries.get_many_mut([
    "Athenæum",
    "Library of Congress",
]);
assert_eq!(
    got,
    Some([
        &mut 1807,
        &mut 1800,
    ]),
);

// Missing keys result in None
let got = libraries.get_many_mut([
    "Athenæum",
    "New York Public Library",
]);
assert_eq!(got, None);

// Duplicate keys result in None
let got = libraries.get_many_mut([
    "Athenæum",
    "Athenæum",
]);
assert_eq!(got, None);

pub unsafe fn get_many_unchecked_mut<Q, const M: usize>( &mut self, ks: [&Q; M] ) -> Option<[&mut V; M]>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Available on crate feature many_mut only.

Attempts to get mutable references to N values in the map at once, without validating that the values are unique.

Returns an array of length N with the results of each query. None will be returned if any of the keys are missing.

For a safe alternative see get_many_mut.

Safety

Calling this method with overlapping keys is undefined behavior even if the resulting references are not used.

Examples

use compact_map::CompactMap;

let mut libraries = CompactMap::default();
libraries.insert("Bodleian Library".to_string(), 1602);
libraries.insert("Athenæum".to_string(), 1807);
libraries.insert("Herzogin-Anna-Amalia-Bibliothek".to_string(), 1691);
libraries.insert("Library of Congress".to_string(), 1800);

let got = libraries.get_many_mut([
    "Athenæum",
    "Library of Congress",
]);
assert_eq!(
    got,
    Some([
        &mut 1807,
        &mut 1800,
    ]),
);

// Missing keys result in None
let got = libraries.get_many_mut([
    "Athenæum",
    "New York Public Library",
]);
assert_eq!(got, None);

pub fn contains_key<Q>(&self, k: &Q) -> bool
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
assert_eq!(map.contains_key(&1), true);
assert_eq!(map.contains_key(&2), false);

pub fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
if let Some(x) = map.get_mut(&1) {
    *x = "b";
}
assert_eq!(map[&1], "b");

pub fn insert(&mut self, k: K, v: V) -> Option<V>

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical. See the [module-level documentation] for more.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
assert_eq!(map.insert(37, "a"), None);
assert_eq!(map.is_empty(), false);

map.insert(37, "b");
assert_eq!(map.insert(37, "c"), Some("b"));
assert_eq!(map[&37], "c");

pub fn try_insert( &mut self, key: K, value: V ) -> Result<&mut V, OccupiedError<'_, K, V, N>>

Available on crate feature map_try_insert only.

Tries to insert a key-value pair into the map, and returns a mutable reference to the value in the entry.

If the map already had this key present, nothing is updated, and an error containing the occupied entry and the value is returned.

Examples

Basic usage:

use compact_map::CompactMap;

let mut map = CompactMap::default();
assert_eq!(map.try_insert(37, "a").unwrap(), &"a");

let err = map.try_insert(37, "b").unwrap_err();
assert_eq!(err.entry.key(), &37);
assert_eq!(err.entry.get(), &"a");
assert_eq!(err.value, "b");

pub fn remove<Q>(&mut self, k: &Q) -> Option<V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
assert_eq!(map.remove(&1), Some("a"));
assert_eq!(map.remove(&1), None);

pub fn remove_entry<Q>(&mut self, k: &Q) -> Option<(K, V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Removes a key from the map, returning the stored key and value if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use compact_map::CompactMap;

let mut map = CompactMap::default();
map.insert(1, "a");
assert_eq!(map.remove_entry(&1), Some((1, "a")));
assert_eq!(map.remove(&1), None);

pub fn into_hashmap(self) -> HashMap<K, V>

Converts the map into a HashMap.

If the map has spilled into a HashMap, this will return that HashMap. Otherwise, it will create a new HashMap and move all the entries into it.

pub fn into_hashmap_with_hasher<S: BuildHasher>( self, hash_builder: S ) -> HashMap<K, V, S>

Converts the map into a HashMap with a given hasher.

This will always create a new HashMap and move all the entries into it.

See also HashMap::with_hasher.

pub fn into_hashmap_with_capacity_and_hasher<S: BuildHasher>( self, capacity: usize, hash_builder: S ) -> HashMap<K, V, S>

Converts the map into a HashMap with at least the specified capacity, using hasher to hash the keys. The capacity will always be at least self.len().

This will always create a new HashMap and move all the entries into it.

See also HashMap::with_capacity_and_hasher.